优化图片显示

+++ app/src/main/java/com/quduo/welfareshop/ui/welfare/entity/NovelInfo.java
@@ -38,7 +38,7 @@ public class NovelInfo extends BaseBean {
     private String thumb_shu;
     private int view_times;
     private int favor_times;
-    private int score;
+    private double score;
     private String description;
 
     public int getId() {
@@ -81,11 +81,11 @@ public class NovelInfo extends BaseBean {
         this.favor_times = favor_times;
     }
 
-    public int getScore() {
+    public double getScore() {
         return score;
     }
 
-    public void setScore(int score) {
+    public void setScore(double score) {
         this.score = score;
     }
